Refinement to the Python-Requires format.
The ~= thing is probably better than just >=3.0; Not that we are close
to Python 4, but let's give the example.

Note also that while PyPI-legacy is lenient on Python-Requires,
warehouse is not (Eg, `3` is interpreted as `>=3.0,<4` for PyPI,
and plainly refused to be uploaded on Warehouse.

Pep 345 is also partly superseded by 440, but this is not indicated,
and the Pep are inconsistent with each other (for now), though it seem
pep 440 should be the winner in case of contradiction.