Grammar and Style Edits to APE0.rst
I worked from the perspective that I did not want to make any changes to the substance of these policies. If anyone feels that any of these edits change the meaning of any clause we should review thoroughly. The only change I did make was to remove the term "ByLaws" and replace it with "policies"  throughout the document. (Though this should not have much of an effect on the substance) This term cannot be used because it is a term with legal connotations for NumFOCUS.  "policies" can be changed to some other term if the community desires.